+++ katana/middleware.py
@@ -67,18 +67,21 @@ class MiddlewareServer(ComponentServer):
 
         code, text = payload.get('response/status').split(' ', 1)
         return {
+            'version': payload.get('response/version', '1.1'),
+            'headers': MultiDict(payload.get('response/headers', {})),
             'status_code': int(code),
             'status_text': text,
             'body': payload.get('response/body', ''),
             }
 
-    def _create_request_component_instance(self, payload):
+    def _create_request_component_instance(self, payload, extra):
         return Request(
             self.__component,
             self.source_file,
             self.component_name,
             self.component_version,
             self.framework_version,
+            attributes=extra.get('attributes'),
             debug=self.debug,
             variables=self.variables,
             # TODO: Use meta and call as arguments instead these many kwargs
@@ -86,13 +89,15 @@ class MiddlewareServer(ComponentServer):
             service_version=payload.get('call/version'),
             action_name=payload.get('call/action'),
             params=payload.get('call/params', []),
+            rid=payload.get('meta/id'),
+            timestamp=payload.get('meta/datetime'),
             gateway_protocol=payload.get('meta/protocol'),
             gateway_addresses=payload.get('meta/gateway'),
             client_address=payload.get('meta/client'),
             http_request=self.http_request_from_payload(payload),
             )
 
-    def _create_response_component_instance(self, payload):
+    def _create_response_component_instance(self, payload, extra):
         return Response(
             Transport(payload.get('transport')),
             self.__component,
@@ -100,6 +105,7 @@ class MiddlewareServer(ComponentServer):
             self.component_name,
             self.component_version,
             self.framework_version,
+            attributes=extra.get('attributes'),
             debug=self.debug,
             variables=self.variables,
             return_value=payload.get('return', NO_RETURN_VALUE),
@@ -110,24 +116,32 @@ class MiddlewareServer(ComponentServer):
             http_response=self.http_response_from_payload(payload),
             )
 
-    def create_component_instance(self, action, payload):
+    def create_component_instance(self, action, payload, extra):
         """Create a component instance for current command payload.
 
         :param action: Name of action that must process payload.
         :type action: str
         :param payload: Command payload.
         :type payload: `CommandPayload`
+        :param extra: A payload to add extra command reply values to result.
+        :type extra: Payload
 
         :rtype: `Request` or `Response`
 
         """
 
         payload = Payload(payload.get('command/arguments'))
+
+        # Always create a new dictionary to store request attributes and save
+        # it inside the command reply extra result values.
+        # If attributes doesn't exist in payload meta use an empty dictionary.
+        extra.set('attributes', dict(payload.get('meta/attributes', {})))
+
         middleware_type = payload.get('meta/type')
         if middleware_type == REQUEST_MIDDLEWARE:
-            return self._create_request_component_instance(payload)
+            return self._create_request_component_instance(payload, extra)
         elif middleware_type == RESPONSE_MIDDLEWARE:
-            return self._create_response_component_instance(payload)
+            return self._create_response_component_instance(payload, extra)
 
     def component_to_payload(self, payload, component):
         """Convert component to a command result payload.
@@ -161,7 +175,7 @@ class MiddlewareServer(ComponentServer):
                 version=http_response.get_protocol_version(),
                 status=http_response.get_status(),
                 body=http_response.get_body(),
-                headers=dict(http_response.get_headers()),
+                headers=dict(http_response.get_headers_array()),
                 )
         else:
             LOG.error('Invalid Middleware callback result')